Use .svelte.md instead of .svx to use Prettier
.svx files cannot be picked by Prettier, so things like wrapping the max
line of the Markdown parts cannot be done. By using .svelte.md, we can
clearly indicate that this is a Svelte + Markdown file, but at least
Prettier treats the file as a regular Markdown file.

If as a side effect, I can remove "mdsvex" from the list of programming
languages in use in this project, it will be good, but it's not a
priority.